% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexmap-old.R
\name{add_hexmap}
\alias{add_hexmap}
\title{Convert map object into hex map}
\usage{
add_hexmap(x, name = "hex", select = "geometry")
}
\arguments{
\item{x}{An sf map object.}

\item{name}{The name of the column that will contain the hex geometry.
By default this is "hex".}

\item{geometry}{A string indicating the name of the}
}
\description{
Convert map object into hex map
}
