% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_apportionment.R
\docType{data}
\name{cds_per_state}
\alias{cds_per_state}
\title{Number of districts per state by decade, 1960-2024}
\format{
A tibble of 50 rows
\describe{
\item{state}{State full name}
\item{st}{State two-character abbrevation}
\item{a1962, a1972, a2022}{Number of congressional districts for the congres
starting in the year 1962, 1972, 2022, etc..}
}
}
\source{
\url{https://en.wikipedia.org/wiki/United_States_Electoral_College}
}
\usage{
cds_per_state
}
\description{
Number of districts each state should have
}
\keyword{datasets}
