% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-tile.R
\name{hex_grid}
\alias{hex_grid}
\title{Create a hexagon grid that tiles a geometry}
\usage{
hex_grid(object, n_tiles = 100)
}
\arguments{
\item{object}{An sf object or geometry to tile. It is recommended to first simplify this geometry to speed up the hexagon tiling.}

\item{n_tiles}{The number of tiles to use for representing the geometry.}
}
\value{
A vector of hexagon geometries that tile the input geometry.
}
\description{
This function will tile a geometry using a specified number of hexagons.
}
\examples{
library(ozmaps)
grid <- hex_grid(abs_ste, n_tiles = 75)

library(ggplot2)
ggplot() +
  geom_sf(data = abs_ste) +
  geom_sf(data = grid, fill = "steelblue", alpha = 0.2)

}
