% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexmap-package.R
\docType{package}
\name{hexmap-package}
\alias{hexmap}
\alias{hexmap-package}
\title{hexmap: Convert a map into a hex map}
\description{
This package takes an sf map object and converts into a hex tile map. This aids data visualisation of maps by making small geographical regions to be equally visible as large geographical regions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/hexmap}
  \item Report bugs at \url{https://github.com/emitanaka/hexmap/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
