% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hexmap.R
\name{make_hex_grid}
\alias{make_hex_grid}
\title{Partitions state geometry into hexagonal units based on CD geometries}
\usage{
make_hex_grid(shp, outline, hex_per_district = 5, infl = 1.05)
}
\description{
Partitions state geometry into hexagonal units based on CD geometries
}
\details{
Usually called by \code{make_hex_map}

@examples
}
\section{New Hampshire in 2005}{
\dontrun{
library(tidyverse)
library(geomander)

cd_geom <- geomander::get_lewis(state = "NH", congress = 109) |>
    mutate(state = "NH", .before = 1) |>
    st_make_valid() |>
    pull(geometry)

st_geom <- summarize(cd_use, geometry = st_union(geometry), .by = state) |>
    pull(geometry)

out <- make_hex_grid(shp = cd_geom, outline = st_geom)

# Show output
ggplot(out$hex) +
    geom_sf() +
    geom_point(data = out$distr, aes(x = X, y = Y))
}
}

