% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hexmap.R
\name{make_hex_map}
\alias{make_hex_map}
\title{Make hex map}
\usage{
make_hex_map(state, d_2020, d_usa, hex_per_district = 5)
}
\arguments{
\item{state}{Character for state}

\item{d_2020}{Dataframe with district labels. Must be a sf shapefile}

\item{d_usa}{State borders. See example}
}
\description{
Main function. Calls \code{make_hex_grid} and \code{place_districts} internally.
}
\examples{
library(dplyr)
library(tidyverse)

## shapefile with 435 rows, one for each CD
cd_shp <- read_rds("data-out/districts_2020_alarm.rds") |>
    filter(!state \%in\% c("AK", "HI"))

st_shp <- cd_shp |>
    summarize(geometry = sf::st_union(geometry), .by = state)

## New Hampshire example
out <- make_hex_map(cd_shp, state = "NH", d_usa = st_shp)

}
