% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-grid.R
\name{map_grid}
\alias{map_grid}
\title{Create a hexagonal grid from geospatial polygons}
\usage{
map_grid(data)
}
\arguments{
\item{data}{An \code{sf} object.}
}
\description{
Create a hexagonal grid from geospatial polygons
}
