% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-match.R
\name{map_match}
\alias{map_match}
\title{Match geospatial polygons}
\usage{
map_match(x, y)
}
\arguments{
\item{x, y}{An \code{sf} object of the same size.}
}
\value{
An sf object with the geometry from \code{x} with rows
matched from \code{y}.
}
\description{
Given two geospatial polygons \code{x} and \code{y}, assign the regions
from \code{y} to the regions in \code{x}.
}
