% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hexmap.R
\name{place_districts}
\alias{place_districts}
\title{Uses the Hungarian algorithm to group (hex) units into \code{n_distr} groups}
\usage{
place_districts(
  res,
  n_runs = 50L,
  max_bursts = 300 + round(sqrt(res$n_distr) * 25),
  silent = FALSE
)
}
\arguments{
\item{res}{Output of \code{make_hex_grid}}
}
\description{
Uses the Hungarian algorithm to group (hex) units into \code{n_distr} groups
}
