% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-allocate.R
\name{tile_allocate}
\alias{tile_allocate}
\title{Allocates tiled geometries to original map geometries}
\usage{
tile_allocate(object, tile)
}
\arguments{
\item{object}{An sf object or geometry that was used to be tiled.}

\item{tile}{An sf object produced with the \code{hex_grid()} function.}
}
\value{
An sf dataset of hexagon geometries with their original data.
}
\description{
This function will allocate tiles to their original geometries. This currently
requires a 1:1 mapping of geometries to their tiles.
}
\examples{
library(ozmaps)
library(dplyr)
australia <- abs_ste \%>\% filter(NAME != "Other Territories")
grid <- hex_grid(australia, n_tiles = 8)
grid <- tile_allocate(australia, grid)

library(ggplot2)
ggplot() +
  geom_sf(data = australia) +
  geom_sf(data = grid, aes(fill = NAME), alpha = 0.2)

}
