% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexmap.R
\name{transition_map_data}
\alias{transition_map_data}
\title{Map transition for animation}
\usage{
transition_map_data(x, map_from, map_to, name = "geometry")
}
\arguments{
\item{x}{An sf map object with two maps.}

\item{map_from, map_to}{Unquoted name of the column with the map to
transition from/to.}

\item{name}{The name of the column with the map object.}
}
\description{
This function transforms the data in the long form so that the transition
from the first map to the second map can be seen in the animation. This function
is like \code{tidyr::pivot_longer} but works for sf objects (error is given
\code{tidyr::pivot_longer} when combining \code{sf} objects).
}
